Lap around Conversational AI using Microsoft Azure Bot & Cognitive Services
Enterprises have a significant need to deliver a conversational assistant tailored to their brand, personalized to their users, and made available across a broad range of devices and social canvases. In this session we will take a dive into building conversational AI solutions leveraging the latest tools from the Azure Bot Service and Cognitive Services. We will also see how to build an end-to-end solution that provides developers and enterprises a set of core foundational capabilities and full control over the end-user experience.